requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
psycopg2-binary==2.9.11
openpyxl==3.1.5
sqlparse==0.5.5
asgiref==3.11.1
